REFACTOR: Consolidate documentation widgets in `help` directory.
